!Impl digital differential analyzer
✓Checker ucc -br
